requirements.txt
Flask
Flask_SQLAlchemy
bcrypt
scikit-learn
pandas
nltk
sentence-transformers
